requirements.txt
Flask==2.3.3
firebase-admin==6.2.0
pandas==2.1.1
python-dotenv==1.0.0
gunicorn==21.2.0 